challenge completed

+++ src/main/java/org/example/service/AcademicRecordService.java
@@ -3,5 +3,5 @@ package org.example.service;
 public interface AcademicRecordService {
   Double calculateAverage();
 
-  Double sumNumberOfGrades();
+  Integer sumNumberOfGrades();
 }
